update Typing module
