+++ compiler/Plato/TypToCore.hs
@@ -19,7 +19,7 @@ import Plato.Driver.Logger
 import Plato.Driver.Monad
 import Plato.Syntax.Core qualified as C
 import Plato.Syntax.Typing qualified as T
-import Plato.Typing.Utils
+import Plato.Typing.Env
 
 elabExpr :: (MonadReader ctx m, HasUniq ctx, MonadIO m) => T.Expr 'T.Typed -> m C.Term
 elabExpr (T.VarE id) = return $ C.Var id
